Refactor demo package structure and add empty config demo.
