+++ Client-PC/src/main/java/uk/ac/cam/cl/wildpetscience/triton/configdemo/ConfigDemo.java
@@ -1,0 +1,9 @@
+package uk.ac.cam.cl.wildpetscience.triton.configdemo;
+
+public class ConfigDemo {
+
+    public static void main(String[] args) {
+
+    }
+
+}
